ascii-video-backend/requirements.txt
flask
opencv-python
numpy
ffmpeg-python
gunicorn
